# PF-BACKEND/swagger/swagger.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Swagger Restaurant",
    "description": "This is a sample Restaurant Store Server",
    "version": "1.0.0"
  },
...
  "paths": {
    "/restaurants": {
      "get": {
        "tags": ["restaurants"],
        "description": "Utilizando esta ruta el servidor respondera con un estatus 200 y enviara la información de todos los restaurantes como un array de objetos paginado, si no se adiciona query de paginado devolvera la pagina 1 por defecto",
        "operationId": "getRestaurants",
        "summary": "Traer todos los restaurantes paginados",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Valor de la pagina para que se muestre las comidas paginadas, en total 4 por cada pagina",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "default": 1,
              "enum": [1, 2, 3]
            }
          },
          {
            "name": "order",
            "in": "query",
            "description": "Organiza ya sea ascendentemente como descendentemente los restaurantes por A-Z y Z-A, tambien se puede hacer por rating de mejor a peor o de peor a mejor, si no se coloca nada te organiza de mayor rating a menor rating",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "enum": ["alphaasc", "alphadesc", "ratingasc", "ratingdesc"]
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Filtra segun las tags que le envias por query, cada tag debe venir separada por el signo '~'",
            "required": false,
            "explode": true,
            "schema": {
              "type": "string",
              "enum": ["pasta", "hamburguesas", "papas"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurants"
                }
              }
            }
          },
          "400": {
            "description": "Ruta /restaurants con parametros invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurants404"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["restaurants"],
        "summary": "Creacion de un restaurante",
        "description": "Esta ruta tomara todos los datos del restaurante por body y lo creara en la base de datos",
        "operationId": "postRestaurants",
        "requestBody": {
          "description": "Ejemplo de lo que espera la ruta por body (no colocar id ni rating)",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Restaurants"
              }
            }
          },
          "readOnly": true,
          "x-displayName": "Post Restaurants"
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurants"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurants404"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Restaurants": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 100000
          },
          "name": {
            "type": "string",
            "example": "Don Julio y Tegui"
          },
          "image": {
            "type": "Buffer",
            "example": "<Buffer 48 65 6c 6c 6f>"
          },
          "type_customer": {
            "type": "string",
            "example": "XXXXX"
          },
          "description": {
            "type": "string",
            "example": "La parrilla Don Julio, ubicada en el barrio de Palermo ..."
          },
          "tags": {
            "type": "array",
            "xml": {
              "XXXX": "xxxx",
              "XX": "xx"
            },
            "items": {}
          },
          "city": {
            "type": "string",
            "example": "Buenos Aires"
          },
          "address": {
            "type": "string",
            "example": "Guatemala 4699 Palermo Viejo"
          },
          "country": {
            "type": "string",
            "example": "Argentina"
          },
          "phoneNumber": {
            "type": "string",
            "example": "+54 11 5311-5668"
          },
          "valoraciones": {
            "type": "array",
            "xml": {
              "XXXX": "xxxx",
              "XX": "xx"
            },
            "items": {}
          },
          "rating": {
            "type": "number",
            "example": 2
          }
        },
        "xml": {
          "name": "restaurants"
        }
      },
      "Restaurants404": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "cannot get invalid route"
          }
        }
      },
...
    }
  }
}